Move `relay::server` module into `relay::server::actor` and move `relay::server::service` to `relay::server`
