+++ iroh-net/src/bin/iroh-relay.rs
@@ -12,7 +12,7 @@ use iroh_net::defaults::{
     DEFAULT_HTTPS_PORT, DEFAULT_HTTP_PORT, DEFAULT_METRICS_PORT, DEFAULT_STUN_PORT,
 };
 use iroh_net::key::SecretKey;
-use iroh_net::relay::iroh_relay;
+use iroh_net::relay;
 use serde::{Deserialize, Serialize};
 use serde_with::{serde_as, DisplayFromStr};
 use tokio_rustls_acme::{caches::DirCache, AcmeConfig};
@@ -356,7 +356,7 @@ async fn main() -> Result<()> {
     let relay_config = build_relay_config(cfg).await?;
     debug!("{relay_config:#?}");
 
-    let mut relay = iroh_relay::Server::spawn(relay_config).await?;
+    let mut relay = relay::server::Server::spawn(relay_config).await?;
 
     tokio::select! {
         biased;
@@ -367,8 +367,8 @@ async fn main() -> Result<()> {
     relay.shutdown().await
 }
 
-/// Convert the TOML-loaded config to the [`iroh_relay::RelayConfig`] format.
-async fn build_relay_config(cfg: Config) -> Result<iroh_relay::ServerConfig<std::io::Error>> {
+/// Convert the TOML-loaded config to the [`relay::server::RelayConfig`] format.
+async fn build_relay_config(cfg: Config) -> Result<relay::server::ServerConfig<std::io::Error>> {
     let tls = match cfg.tls {
         Some(ref tls) => {
             let cert_config = match tls.cert_mode {
@@ -382,7 +382,7 @@ async fn build_relay_config(cfg: Config) -> Result<iroh_relay::ServerConfig<std:
                         anyhow::Ok((key, certs))
                     })
                     .await??;
-                    iroh_relay::CertConfig::Manual { private_key, certs }
+                    relay::server::CertConfig::Manual { private_key, certs }
                 }
                 CertMode::LetsEncrypt => {
                     let hostname = tls
@@ -397,17 +397,17 @@ async fn build_relay_config(cfg: Config) -> Result<iroh_relay::ServerConfig<std:
                         .contact([format!("mailto:{}", contact)])
                         .cache_option(Some(DirCache::new(tls.cert_dir())))
                         .directory_lets_encrypt(tls.prod_tls);
-                    iroh_relay::CertConfig::LetsEncrypt { config }
+                    relay::server::CertConfig::LetsEncrypt { config }
                 }
             };
-            Some(iroh_relay::TlsConfig {
+            Some(relay::server::TlsConfig {
                 https_bind_addr: tls.https_bind_addr(&cfg),
                 cert: cert_config,
             })
         }
         None => None,
     };
-    let limits = iroh_relay::Limits {
+    let limits = relay::server::Limits {
         accept_conn_limit: cfg
             .limits
             .as_ref()
@@ -419,16 +419,16 @@ async fn build_relay_config(cfg: Config) -> Result<iroh_relay::ServerConfig<std:
             .map(|l| l.accept_conn_burst)
             .unwrap_or_default(),
     };
-    let relay_config = iroh_relay::RelayConfig {
+    let relay_config = relay::server::RelayConfig {
         secret_key: cfg.secret_key.clone(),
         http_bind_addr: cfg.http_bind_addr(),
         tls,
         limits,
     };
-    let stun_config = iroh_relay::StunConfig {
+    let stun_config = relay::server::StunConfig {
         bind_addr: cfg.stun_bind_addr(),
     };
-    Ok(iroh_relay::ServerConfig {
+    Ok(relay::server::ServerConfig {
         relay: Some(relay_config),
         stun: Some(stun_config),
         #[cfg(feature = "metrics")]
